requirements.txt
ConfigArgParse~=1.4
black
prospector
discord.py
requests~=2.25.1
dataset~=1.4.5
APScheduler~=3.7.0